// apps/dashboard/src/types/enums.ts
/* eslint-disable no-unused-vars */
export enum APITags {
  ContactById = 'ContactById',
  ContactList = 'ContactList',
  CityList = 'CityList',
  AdvertisementList = 'AdvertisementList',
  ApproachList = 'ApproachList',
  FindUsList = 'FindUsList',
  GalleryList = 'GalleryList',
  CategoryList = 'CategoryList',
  SubcategoryList = 'SubcategoryList',
  JumbotronList = 'JumbotronList',
  WelcomeList = 'WelcomeList',
  ReviewList = 'ReviewList',
  HomeList = 'HomeList',
  RequestList = 'RequestList',
  UserList = 'UserList',
}

export enum APIMethods {
  POST = 'POST',
  GET = 'GET',
  PUT = 'PUT',
  DELETE = 'DELETE',
  PATCH = 'PATCH',
}

export enum DashboardPaths {
  ROOT = '/',
  ADVERTISEMENT = '/advertisement',
  ADVERTISEMENT_DETAIL = '/advertisement-detail',
  APPROACH = '/approach',
  APPROACH_DETAIL = '/approach-detail',
  CITY = '/city',
  CITY_DETAIL = '/city-detail',
  CONTACT = '/contact',
  CONTACT_DETAIL = '/contact-detail',
  FIND_US = '/find-us',
  FIND_US_DETAIL = '/find-us-detail',
  GALLERY = '/gallery',
  GALLERY_DETAIL = '/gallery-detail',
  CATEGORY = '/category',
  CATEGORY_DETAIL = '/category-detail',
  SUBCATEGORY = '/subcategory',
  SUBCATEGORY_DETAIL = '/subcategory-detail',
  JUMBOTRON = '/jumbotron',
  JUMBOTRON_DETAIL = '/jumbotron-detail',
  WELCOME = '/welcome',
  WELCOME_DETAIL = '/welcome-detail',
  REVIEW = '/review',
  REVIEW_DETAIL = '/review-detail',
  HOME = '/home',
  HOME_DETAIL = '/home-detail',
  REQUEST = '/request',
  REQUEST_DETAIL = '/request-detail',
  USER = '/user',
  USER_DETAIL = '/user-detail',
}